requirements-local.txt
# Local development requirements (for running scripts/quick_dump.py and smoke_run.py)
playwright==1.49.1
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.3
tenacity==9.0.0
psycopg2-binary==2.9.9
boto3==1.35.72
httpx==0.27.2
python-dotenv==1.0.0
duckdb==1.1.0

